Added lab 1 files.
Added the files necessary to lab 1. Extracted the .zip's that Alex
provided. It'll only take up about 13M according to du. Hopefully this
will make testing with our scripts easier.

It may be necessary to pare these out in the future. This is a big data
class and out data may get too out of control for github to allow.