vocabulary data structure with crucial features
Built a vocabulary object that implements the "bag of words" schema we learned in class. The bag is a python dict/hasttable, with an additional word list that remains sorted as more words are added.

An iterator on a Vocabulary object is really simple, since it simply returns the iterator for the sorted list. This can be used to print out an alphabetical list of the words in the bag.